fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Jinja2==3.1.4
passlib==1.7.4
python-jose[cryptography]==3.3.0
pytest==8.3.3
